+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - #362 - add option for `Last Login`, where we can use `never` or we can use date as `__current_date__` where it is replaced by the current date or `__current_date__@-5 days` where it takes the date for 5 days ago. It only can be used with days + and - days after `@`
  - #331 - add fields `min` and `max` for characters size in style `textarea`
  - #331 - add fields `min` and `max` for characters size in style `input`
+ - #330 - check the user locale and if the locale exists it is loaded as default for request
 
 # v6.4.9
 ### Bugfix