+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v6.5.0 (Not published)
+# v6.5.0
 ### New features
  - move data config fetch in `BaseModel` from `StyleModel` 
  - #362 - add option for `Last Login`, where we can use `never` or we can use date as `__current_date__` where it is replaced by the current date or `__current_date__@-5 days` where it takes the date for 5 days ago. It only can be used with days + and - days after `@`